cleaner/requirements.txt
pandas==2.2.2
numpy==1.26.4
duckdb==1.0.0
minio==7.2.7
pika==1.3.2
python-dotenv==1.0.1
prometheus-client==0.20.0
